Creates simulations of group sequential trial data and calculates test statistic based on user-defined stat function.

Creates simulations of group sequential trial data and calculates test statistic based on user-defined stat function.

create.simulations(n_sims, n_k, stat.func, data.generator, ...)

Arguments

n_sims
Number of simulations

n_k
Sample size per stage

stat.func
Function (or list of functions) to calculate the test statistic based on the generated data, at each observed information rates. If this is a list of functions, it needs to be the same length as the sequence of boundaries.

data.generator
Function to generate data. The output of this function should be a matrix, which is the input for stat.func. First argument of this function needs to be n, the sample size, and the function needs to be vectorized.

...
Additional arguments to be passed to data.generator

Examples

create.simulations(10, n_k=c(10, 10, 10),
                   stat.func=list(mean, sum, mean),
                   data.generator=function(n) as.matrix(rnorm(n)))

#>              [,1]       [,2]        [,3]
#>  [1,] -0.44136451 -6.2006044 -0.24787308
#>  [2,]  0.15216644  2.6820571  0.13179875
#>  [3,] -0.08901941  1.9181066  0.14306822
#>  [4,]  0.62691039  0.9077428  0.06248182
#>  [5,]  0.44419281  2.2481136  0.18737804
#>  [6,]  0.18548195 -2.2718804  0.02814191
#>  [7,]  0.48246965  9.4147306  0.21583993
#>  [8,] -0.08981122 -1.8656430 -0.16760277
#>  [9,]  0.31660067  1.1790747 -0.07576693
#> [10,]  0.30906306  5.8739720  0.24119612