Derive boundaries with alpha-spending.

Get the boundaries required for a particular alpha-spending function and an observed information rate out of a total sample size.

get.boundaries.aspend(
  a.func,
  a,
  rates,
  N,
  n_sims,
  stat.func,
  data.generator,
  ...
)

Arguments

a.func
A continuous, monotonic increasing function of t where a.func(a, t=0) = 0 and a.func(a, t=1) = a where a is the type I error desired

a
The type I error desired

rates
A vector of information rates (between 0 and 1)

N
maximum total sample size

n_sims
Number of Monte-Carlo simulations

stat.func
Function (or list of functions) to calculate the test statistic based on the generated data, at each observed information rates. If this is a list of functions, it needs to be the same length as the sequence of boundaries.

data.generator
Function to generate data. The output of this function should be a matrix, which is the input for stat.func. First argument of this function needs to be n, the sample size, and the function needs to be vectorized.

Details

Your data generator in this case needs to be the data generation under the null hypothesis in order for this function to work correctly with alpha-spending.

Examples

set.seed(101)
# information rates
t <- 1:3/3

# approximate Pocock boundaries
a.func.pocock <- function(a, t) a * log(1 + (exp(1) - 1) * t)
get.boundaries.aspend(a.func=a.func.pocock, a=0.05,
                      rates=t, N=1000, n_sims=1000,
                      stat.func=function(x) mean(x) * sqrt(length(x)),
                      data.generator=function(n) rnorm(n))

#> Error in data[1:x, ]: incorrect number of dimensions

# approximate O'Brien-Fleming boundaries
a.func.obf <- function(a, t) 4 * (1 - pnorm(qnorm(1-a/4)/sqrt(t)))
get.boundaries.aspend(a.func=a.func.obf, a=0.05,
                      rates=t, N=1000, n_sims=1000,
                      stat.func=function(x) mean(x) * sqrt(length(x)),
                      data.generator=function(n) rnorm(n))

#> Error in data[1:x, ]: incorrect number of dimensions